Fix checking of duplicate settings
Currently shoop.apps.settings module tries to check that no setting is
declared in two separate apps.  It should raise ImproperlyConfigured
exception when that happens, but the code was left in broken state on
refactoring and therefore AttributeError was raised instead.

Fix the code to provide more info about which apps are declaring the
conflicting setting.

Refs SHOOP-1277