OrderSource: Use `add_line` for adding new lines
Previous method of adding lines to OrderSource (or BaseBasket) was to
directly modify its lines through `source.lines` list.  This does not
allow the OrderSource to "see" when its lines are being changed.

Remove direct access to lines by renaming the container to `_lines` and
add `add_line` method for adding new line to OrderSource.

Most important reason for this: This allows automatic uncaching on
adding new lines.

Refs SHOOP-975