PricingModule: Make get_context simpler
Remove the `context = context.pricing_context` trick (which was there
for `PriceTaxContext` support) and the final fallback of calling
`get_context_from_data`.

Reason for this is, that the most common way to get a pricing context is
through a request and that makes request somewhat special for
`get_context`, but other cases should be handled explicitly by creating
a `PricingContext` with constructor of a correct subclass or with
`get_context_from_data` method.  This will make code of API users more
explicit and easier to read.

Refs SHOOP-1636/SHOOP-1735