Pricing context for update_order_line_from_product
Make `update_order_line_from_product` to require a PricingContextable
rather than request (which is a PricingContextable too).

This makes `update_order_line_from_product` more general and previously
the reason why a request is needed was not so clear, now it is more
explicit that the reason is pricing.

Refs SHOOP-1636/SHOOP-1735