Add "setup.py build_messages" command
Add a "build_messages" subcommand for setup.py which does "django-admin
compilemessages" for all found locale directories, i.e. compiles
the *.po message catalogs to binary *.mo files.

The "build_messages" command is also added as a dependency to "build"
command so that it will be automatically called on "pip install" or on
"pip wheel" / "setup.py bdist_wheel".

This should be usable for external projects too.

Refs SHOOP-1829/SHOOP-1886