Add test case for invalid choices in EnumIntegerField
With Python 2.7, enum34 1.1 and django-enumfields 0.7.4, the
EnumIntegerField returns invalid choices, because on that setup a zero
valued enum item will evaluate to false and
`EnumIntegerField.get_choices` converts enum items to values only if the
enum item evaluates to true.

A fix for this is pushed upstream to django-enumfields.  See
https://github.com/hzdg/django-enumfields/pull/44 for more info.

This problem was found out while updating the dependencies for
SHOOP-1954.

Refs SHOOP-1942/SHOOP-1954