TranslatableShoopModel: Make str be always non-empty
Make `__str__` of any TranslatableShoopModel always return non-empty
value.  This is good, because e.g. in Admin, when representation of an
object is rendered and there is non translation for some language, the
empty value is misleading.

If the empty value must be rendered for some reason, it is still
accessible through `object.name` (or whatever the name attribute is).

Refs SHOOP-2397