Core: Make `OrderSource.language` usable
Old language was pretty much unused and the value was
only applied if the programmer knew that it should be set.

Change the functionality so that if the `OrderSource`
has no language set, the customers
language will be used as a language for `OrderSource` too.

`OrderSource` must have a language set, so if no valid
languages are found, source will use `settings.LANGUAGE_CODE`.

Also add tests to show that this change makes notification emails to
be send in proper language.

Refs SH-428