Cache all displayed product prices
For variation products also cache priced children since like
99% of time the priced children is used to calculate the
pricing range. We are willing to pay this price (maybe).

Also this makes sure that prices is calculated only once per
request. I had a issues with memcache write speed so that
is the reason why the prices is cached to request.

From 12.3s -> 4.8s