Front: fix manufacturer modified filter to consider only visible shop products
Issue: some manufacturers were available for filtering and when applied, no product was returned.
The queryset was not considering invisible shop products that had the manufacturer linked.
Excluding invisible shop products fixes the issue.

The current shop was also added to the manufacturer and shop product querysets to avoid considering other shops.

Refs POS-2497