Front: bump template helpers cache when products get saved
When a shop product was saved, the context cache for the template helpers were not bumped as there were no item passed to the context cache.

To fix the issue we pass an unique string with the shop ID as the item to make it possible to bump the correct cache namespace.

Refs #1347