Core: improve to_aware function to consider DST cases
When converting from date to datetime the time component will be
always 00:00 and in some cases, after applying the timezone information,
pytz will raise exception because the time can not exists at all.
This will occur when the date is a date which the timezone changes to DST,
adding 1 hour to time at 00:00 o'clock of the previous date.

To fix the issue, we first try making the datetime aware without DST mode,
if pytz raises NonExistentTimeError is because that time doesn't exist,
this way we try again but now informing that DST should be used instead.

No Refs