Fix search results overflowing the canvas
Currently the search-result-canvas always has space on the bottom. But 
if there are more search results than can fit the screen, a visual bug 
happens, with results being displayed, but outside the canvas.

This change makes it so that the canvas will continue until the very 
bottom edge of the screen. Only when needed, otherwise there is still a 
space.